Refactored dummy methods into StudentSummaryDisplay
