+++ src/app/domain/dummyComponentState.ts
@@ -1,0 +1,7 @@
+import { ComponentState } from './componentState';
+
+export class DummyComponentState extends ComponentState {
+  constructor(jsonObject: any = {}) {
+    super(jsonObject);
+  }
+}
