+++ src/assets/wise5/directives/teacher-summary-display/teacher-summary-display.component.ts
@@ -8,6 +8,9 @@ import { TeacherDataService } from '../../services/teacherDataService';
 import { CommonModule } from '@angular/common';
 import { MatCardModule } from '@angular/material/card';
 import { HighchartsChartModule } from 'highcharts-angular';
+import { Observable } from 'rxjs';
+import { Annotation } from '../../common/Annotation';
+import { ComponentState } from '../../../../app/domain/componentState';
 
 @Component({
   imports: [CommonModule, HighchartsChartModule, MatCardModule],
@@ -25,5 +28,27 @@ export class TeacherSummaryDisplayComponent extends SummaryDisplayComponent {
     protected summaryService: SummaryService
   ) {
     super(annotationService, configService, dataService, projectService, summaryService);
+  }
+
+  protected getLatestScores(): Observable<Annotation[]> {
+    return this.getLatestStudentScores();
+  }
+
+  protected getLatestWork(): Observable<ComponentState[]> {
+    return this.getLatestStudentWork();
+  }
+
+  protected renderResponsesOrScores(isRenderingResponses: boolean): void {
+    if (this.isSourceSelf()) {
+      this.displaySourceSelfMessageToTeacher();
+    } else {
+      isRenderingResponses ? this.renderClassResponses() : this.renderClassScores();
+    }
+  }
+
+  private displaySourceSelfMessageToTeacher(): void {
+    this.doRender = false;
+    this.warningMessage = $localize`The student will see a graph of their individual data here.`;
+    this.hasWarning = true;
   }
 }
